requirements.txt
pyaudio
google-genai
google-generativeai
pillow
requests
edge-tts
beautifulsoup4
playwright
pyautogui
pyperclip
opencv-python
mediapipe
numpy
mss
psutil
send2trash
comtypes
pycaw
win10toast
duckduckgo-search
youtube-transcript-api
python-kasa
openpyxl
